Thrift changes
