+++ rubix-core/src/main/java/com/qubole/rubix/core/CachingInputStream.java
@@ -18,8 +18,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
+import com.qubole.rubix.bookkeeper.BookKeeperService;
 import com.qubole.rubix.bookkeeper.Location;
-import com.qubole.rubix.bookkeeper.RetryingBookkeeperClient;
 import com.qubole.rubix.spi.CacheConfig;
 import com.qubole.rubix.spi.ClusterType;
 import org.apache.commons.logging.Log;
@@ -40,7 +40,7 @@ import java.util.concurrent.Executors;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
-import static com.qubole.rubix.bookkeeper.BookKeeperClient.createBookKeeperClient;
+import static com.qubole.rubix.bookkeeper.RetryingBookkeeperClient.createBookKeeperClient;
 
 /**
  * Created by stagra on 29/12/15.
@@ -64,7 +64,7 @@ public class CachingInputStream
     private String localPath;
     private long lastModified;
 
-    private RetryingBookkeeperClient bookKeeperClient;
+    private BookKeeperService.Client bookKeeperClient;
     private Configuration conf;
 
     private boolean strictMode = false;
