+++ rubix-core/src/main/java/com/qubole/rubix/core/RemoteReadRequestChain.java
@@ -13,7 +13,7 @@
 package com.qubole.rubix.core;
 
 import com.google.common.base.Throwables;
-import com.qubole.rubix.bookkeeper.RetryingBookkeeperClient;
+import com.qubole.rubix.bookkeeper.BookKeeperService;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.conf.Configuration;
@@ -25,7 +25,7 @@ import java.nio.MappedByteBuffer;
 import java.nio.channels.FileChannel;
 
 import static com.google.common.base.Preconditions.checkState;
-import static com.qubole.rubix.bookkeeper.BookKeeperClient.createBookKeeperClient;
+import static com.qubole.rubix.bookkeeper.RetryingBookkeeperClient.createBookKeeperClient;
 
 /**
  * Created by stagra on 4/1/16.
@@ -149,7 +149,7 @@ public class RemoteReadRequestChain
     public void updateCacheStatus(String remotePath, long fileSize, long lastModified, int blockSize, Configuration conf)
     {
         try {
-            RetryingBookkeeperClient client = createBookKeeperClient(conf);
+            BookKeeperService.Client client = createBookKeeperClient(conf);
             for (ReadRequest readRequest : readRequests) {
                 client.setAllCached(remotePath, fileSize, lastModified, toBlock(readRequest.getBackendReadStart(), blockSize), toBlock(readRequest.getBackendReadEnd() - 1, blockSize) + 1);
             }
